requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
groq
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
pydantic